fix: instance vs type confusion
Fix an issue where we weren't instantiating the strategy in the current default demo.

This motivated us passing a superfluous 'self' elsewhere (in the play method of BotPlayer).

The tests were doing the correct thing, and thus failed, so we had a fun situations where making one pass would cause the other to fail (judging by the surface level errors alone).

It turns out it was really all as simple as a couple of sets of parens.