ci: move and disable lint job for now
Moving to the main PR workflow, so that this will eventually be run as part of the checks for PRs.

It seems to be complaining about a number of things, and doesn't pass at the moment, so it doesn't make sense to run it regularly yet.

Will re-enable once it has a chance of passing.